+++ packages/EventMap/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.1] 2021-09-22
+### Updated
+- dependencies and audit
+
+
+## [0.7.0] 2021-09-22
+### Breaking Change
+- Moved decorators to a decorators file
+
 ## [0.6.1] - 2021-09-10
 ### Updated
 - Documentation